Revert "move anonymous pointcloud functions into their respective utils files…"
This reverts commit 5808f65f3ded0d2460c11e948d30026aae4621e1.
