+++ multisense_ros/include/multisense_ros/camera_utilities.h
@@ -146,7 +146,6 @@ sensor_msgs::CameraInfo makeCameraInfo(const crl::multisense::image::Config& con
 RectificationRemapT makeRectificationRemap(const crl::multisense::image::Config& config,
                                            const crl::multisense::image::Calibration::Data& calibration,
                                            const crl::multisense::system::DeviceInfo& device_info);
-
 class StereoCalibrationManager
 {
 public:
@@ -186,16 +185,11 @@ public:
     ///
     /// @brief Reproject disparity values into 3D
     ///
-    static Eigen::Vector3f reproject(size_t u,
-                                     size_t v,
-                                     double d,
-                                     const sensor_msgs::CameraInfo &left_camera_info,
-                                     const sensor_msgs::CameraInfo &right_camera_info);
-
-    ///
-    /// @brief Check if a reprojected point falls within a valid range
-    ///
-    static bool isValidReprojectedPoint(const Eigen::Vector3f& pt, const float squared_max_range);
+    Eigen::Vector3f reproject(size_t u,
+                              size_t v,
+                              double d,
+                              const sensor_msgs::CameraInfo &left_camera_info,
+                              const sensor_msgs::CameraInfo &right_camera_info) const;
 
     ///
     /// @brief Project points corresponding to disparity measurements in the left rectified image frame into the
@@ -207,8 +201,8 @@ public:
     /// @brief Project points corresponding to disparity measurements in the left rectified image frame into the
     ///        aux rectified image plane
     ///
-    static Eigen::Vector2f rectifiedAuxProject(const Eigen::Vector3f &left_rectified_point,
-                                               const sensor_msgs::CameraInfo &aux_camera_info);
+    Eigen::Vector2f rectifiedAuxProject(const Eigen::Vector3f &left_rectified_point,
+                                        const sensor_msgs::CameraInfo &aux_camera_info) const;
 
     ///
     /// @brief Get the current main stereo pair operating resolution. This resolution applies for both the mono
@@ -268,20 +262,6 @@ private:
     std::shared_ptr<RectificationRemapT> left_remap_;
     std::shared_ptr<RectificationRemapT> right_remap_;
 };
-
-///
-/// @brief Determine if a pixel should be clipped given the provided parameters
-///
-bool clipPoint(const BorderClip& borderClipType,
-               double borderClipValue,
-               size_t height,
-               size_t width,
-               size_t u,
-               size_t v);
-
-///
-/// @brief Given a floating point pixel location, determine its average color using all neighboring pixels
-cv::Vec3b interpolateColor(const Eigen::Vector2f &pixel, const cv::Mat &image);
 
 }// namespace
