+++ multisense_ros/src/camera.cpp
@@ -115,6 +115,135 @@ void groundSurfaceCB(const image::Header& header, void* userDataP)
 void groundSurfaceSplineCB(const ground_surface::Header& header, void* userDataP)
 { reinterpret_cast<Camera*>(userDataP)->groundSurfaceSplineCallback(header); }
 
+
+bool isValidReprojectedPoint(const Eigen::Vector3f& pt, float squared_max_range)
+{
+    return pt[2] > 0.0f && std::isfinite(pt[2]) && pt.squaredNorm() < squared_max_range;
+}
+
+void writePoint(sensor_msgs::PointCloud2 &pointcloud, const size_t index, const Eigen::Vector3f &point)
+{
+    assert(index * pointcloud.point_step < pointcloud.data.size());
+
+    assert(pointcloud.fields.size() >= 3);
+    assert(pointcloud.fields[0].datatype == messageFormat<float>());
+    assert(pointcloud.fields[1].datatype == messageFormat<float>());
+    assert(pointcloud.fields[2].datatype == messageFormat<float>());
+
+    float* cloudP = reinterpret_cast<float*>(&(pointcloud.data[index * pointcloud.point_step]));
+    cloudP[0] = point[0];
+    cloudP[1] = point[1];
+    cloudP[2] = point[2];
+}
+
+template <typename ColorT>
+void writePoint(sensor_msgs::PointCloud2 &pointcloud, const size_t index, const Eigen::Vector3f &point, const ColorT &color)
+{
+    writePoint(pointcloud, index, point);
+
+    assert(pointcloud.fields.size() == 4);
+    assert(pointcloud.fields[3].datatype == messageFormat<ColorT>());
+
+    ColorT* colorP = reinterpret_cast<ColorT*>(&(pointcloud.data[index * pointcloud.point_step + pointcloud.fields[3].offset]));
+    colorP[0] = color;
+}
+
+void writePoint(sensor_msgs::PointCloud2 &pointcloud,
+                const size_t pointcloud_index,
+                const Eigen::Vector3f &point,
+                const size_t image_index,
+                const image::Header &image)
+{
+    switch (image.bitsPerPixel)
+    {
+        case 8:
+        {
+            const uint8_t luma = reinterpret_cast<const uint8_t*>(image.imageDataP)[image_index];
+            return writePoint(pointcloud, pointcloud_index, point, luma);
+        }
+        case 16:
+        {
+            const uint16_t luma = reinterpret_cast<const uint16_t*>(image.imageDataP)[image_index];
+            return writePoint(pointcloud, pointcloud_index, point, luma);
+        }
+        default:
+        {
+            throw std::runtime_error("Invalid bits per pixel value");
+        }
+    }
+}
+
+bool clipPoint(const BorderClip& borderClipType,
+               double borderClipValue,
+               size_t height,
+               size_t width,
+               size_t u,
+               size_t v)
+{
+    switch (borderClipType)
+    {
+        case BorderClip::NONE:
+        {
+            return false;
+        }
+        case BorderClip::RECTANGULAR:
+        {
+            return !( u >= borderClipValue && u <= width - borderClipValue &&
+                      v >= borderClipValue && v <= height - borderClipValue);
+        }
+        case BorderClip::CIRCULAR:
+        {
+            const double halfWidth = static_cast<double>(width)/2.0;
+            const double halfHeight = static_cast<double>(height)/2.0;
+
+            const double radius = sqrt( halfWidth * halfWidth + halfHeight * halfHeight ) - borderClipValue;
+
+            return !(Eigen::Vector2d{halfWidth - u, halfHeight - v}.norm() < radius);
+        }
+        default:
+        {
+            ROS_WARN("Camera: Unknown border clip type.");
+            break;
+        }
+    }
+
+    return true;
+}
+
+cv::Vec3b interpolate_color(const Eigen::Vector2f &pixel, const cv::Mat &image)
+{
+    const float width = image.cols;
+    const float height = image.rows;
+
+    const float &u = pixel(0);
+    const float &v = pixel(1);
+
+    //
+    // Implement a basic bileinar interpolation scheme
+    // https://en.wikipedia.org/wiki/Bilinear_interpolation
+    //
+    const size_t min_u = static_cast<size_t>(std::min(std::max(std::floor(u), 0.f), width - 1.f));
+    const size_t max_u = static_cast<size_t>(std::min(std::max(std::floor(u) + 1, 0.f), width - 1.f));
+    const size_t min_v = static_cast<size_t>(std::min(std::max(std::floor(v), 0.f), height - 1.f));
+    const size_t max_v = static_cast<size_t>(std::min(std::max(std::floor(v) + 1, 0.f), height - 1.f));
+
+    const cv::Vec3d element00 = image.at<cv::Vec3b>(width * min_v + min_u);
+    const cv::Vec3d element01 = image.at<cv::Vec3b>(width * min_v + max_u);
+    const cv::Vec3d element10 = image.at<cv::Vec3b>(width * max_v + min_u);
+    const cv::Vec3d element11 = image.at<cv::Vec3b>(width * max_v + max_u);
+
+    const size_t delta_u = max_u - min_u;
+    const size_t delta_v = max_v - min_v;
+
+    const double u_ratio = delta_u == 0 ? 1. : (static_cast<double>(max_u) - u) / static_cast<double>(delta_u);
+    const double v_ratio = delta_v == 0 ? 1. : (static_cast<double>(max_v) - v) / static_cast<double>(delta_v);
+
+    const cv::Vec3b f_xy0 = element00 * u_ratio + element01 * (1. - u_ratio);
+    const cv::Vec3b f_xy1 = element10 * u_ratio + element11 * (1. - u_ratio);
+
+    return (f_xy0 * v_ratio + f_xy1 * (1. - v_ratio));
+}
+
 } // anonymous
 
 //
@@ -1675,7 +1804,7 @@ void Camera::pointCloudCallback(const image::Header& header)
         rectified_color = std::move(rect_rgb_image);
     }
 
-    // disparity conversion requires valid right calibration
+    // disparity conversion requires valid right image
     if (!stereo_calibration_manager_->validRight())
     {
         throw std::runtime_error("Stereo calibration manager missing right calibration");
@@ -1732,7 +1861,7 @@ void Camera::pointCloudCallback(const image::Header& header)
                 packed_color = 0;
 
                 const auto color_pixel = (has_aux_camera_ && disparity != 0.0) ?
-                    interpolateColor(stereo_calibration_manager_->rectifiedAuxProject(point, aux_camera_info),
+                    interpolate_color(stereo_calibration_manager_->rectifiedAuxProject(point, aux_camera_info),
                                       rectified_color) :
                     rectified_color.at<cv::Vec3b>(y, x);
 
@@ -1748,7 +1877,7 @@ void Camera::pointCloudCallback(const image::Header& header)
             {
                 if (pub_organized_pointcloud)
                 {
-                    writePoint(luma_organized_point_cloud_, index, invalid_point, index, left_luma_rect->data().bitsPerPixel, left_luma_rect->data().imageDataP);
+                    writePoint(luma_organized_point_cloud_, index, invalid_point, index, left_luma_rect->data());
                 }
 
                 if (pub_color_organized_pointcloud)
@@ -1759,11 +1888,11 @@ void Camera::pointCloudCallback(const image::Header& header)
                 continue;
             }
 
-            const bool valid = stereo_calibration_manager_->isValidReprojectedPoint(point, squared_max_range);
+            const bool valid = isValidReprojectedPoint(point, squared_max_range);
 
             if (pub_pointcloud && valid)
             {
-                writePoint(luma_point_cloud_, valid_points, point, index, left_luma_rect->data().bitsPerPixel, left_luma_rect->data().imageDataP);
+                writePoint(luma_point_cloud_, valid_points, point, index, left_luma_rect->data());
             }
 
             if(pub_color_pointcloud && valid)
@@ -1773,7 +1902,7 @@ void Camera::pointCloudCallback(const image::Header& header)
 
             if (pub_organized_pointcloud)
             {
-                writePoint(luma_organized_point_cloud_, index, valid ? point : invalid_point, index, left_luma_rect->data().bitsPerPixel, left_luma_rect->data().imageDataP);
+                writePoint(luma_organized_point_cloud_, index, valid ? point : invalid_point, index, left_luma_rect->data());
             }
 
             if (pub_color_organized_pointcloud)
